pyproject.toml
[tool.poetry]
name = "domjudge-tool-web"
version = "0.1.0"
description = ""
authors = ["jasonxie"]

[tool.poetry.dependencies]
python = "^3.8"
django = "*"
django-extensions = "*"
pyyaml = "*"
django-jazzmin = "*"
django-environ = "*"
django-split-settings = "*"
psycopg2-binary = "*"
sentry-sdk = "*"
django-object-actions = "*"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
drf-yasg = "^1.21.4"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
werkzeug = "*"
flake8 = "*"
flake8-commas = "*"
mypy = "*"
yapf = "*"
ipython = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
